# [opt](column nullable) Optimization for ColumnNullable serialization