--- a/be/src/vec/columns/column_decimal.cpp
+++ b/be/src/vec/columns/column_decimal.cpp
@@ -21,6 +21,7 @@
 #include "vec/columns/column_decimal.h"
 
 #include <fmt/format.h>
+#include <glog/logging.h>
 
 #include <limits>
 #include <ostream>
@@ -37,6 +38,7 @@
 #include "vec/common/sip_hash.h"
 #include "vec/common/unaligned.h"
 #include "vec/core/sort_block.h"
+#include "vec/core/types.h"
 #include "vec/data_types/data_type.h"
 
 template <typename T>
@@ -86,11 +88,29 @@ void ColumnDecimal<T>::serialize_vec(std::vector<StringRef>& keys, size_t num_ro
 
 template <typename T>
 void ColumnDecimal<T>::serialize_vec_with_null_map(std::vector<StringRef>& keys, size_t num_rows,
-                                                   const uint8_t* null_map) const {
-    for (size_t i = 0; i < num_rows; ++i) {
-        if (null_map[i] == 0) {
-            memcpy_fixed<T>(const_cast<char*>(keys[i].data + keys[i].size), (char*)&data[i]);
-            keys[i].size += sizeof(T);
+                                                   const UInt8* null_map) const {
+    DCHECK(null_map != nullptr);
+    const bool has_null = simd::contain_byte(null_map, num_rows, 1);
+    if (has_null) {
+        for (size_t i = 0; i < num_rows; ++i) {
+            char* __restrict dest = const_cast<char*>(keys[i].data + + keys[i].size);
+            // serialize null first
+            memcpy(dest, null_map + i, sizeof(UInt8));
+            if (null_map[i] == 0) {
+                memcpy_fixed<T>(dest + 1, (char*)&data[i]);
+                keys[i].size += sizeof(T) + sizeof(UInt8);
+            } else {
+                keys[i].size += sizeof(UInt8);
+            }
+        }
+    } else {
+        for (size_t i = 0; i < num_rows; ++i) {
+            if (null_map[i] == 0) {
+                char* __restrict dest = const_cast<char*>(keys[i].data + + keys[i].size);
+                memset(dest, 0, 1);
+                memcpy_fixed<T>(dest + 1, (char*)&data[i]);
+                keys[i].size += sizeof(T) + sizeof(UInt8);
+            }
         }
     }
 }