--- a/be/src/vec/columns/column_nullable.cpp
+++ b/be/src/vec/columns/column_nullable.cpp
@@ -245,22 +245,8 @@ size_t ColumnNullable::get_max_row_byte_size() const {
 
 void ColumnNullable::serialize_vec(std::vector<StringRef>& keys, size_t num_rows,
                                    size_t max_row_byte_size) const {
-    if (has_null()) {
-        const auto& arr = get_null_map_data();
-        for (size_t i = 0; i < num_rows; ++i) {
-            auto* val = const_cast<char*>(keys[i].data + keys[i].size);
-            *val = (arr[i] ? 1 : 0);
-            keys[i].size++;
-        }
-        get_nested_column().serialize_vec_with_null_map(keys, num_rows, arr.data());
-    } else {
-        for (size_t i = 0; i < num_rows; ++i) {
-            auto* val = const_cast<char*>(keys[i].data + keys[i].size);
-            *val = 0;
-            keys[i].size++;
-        }
-        get_nested_column().serialize_vec(keys, num_rows, max_row_byte_size);
-    }
+    const auto& arr = get_null_map_data();
+    get_nested_column().serialize_vec_with_null_map(keys, num_rows, arr.data());
 }
 
 void ColumnNullable::deserialize_vec(std::vector<StringRef>& keys, const size_t num_rows) {