--- a/be/src/vec/columns/column_vector.cpp
+++ b/be/src/vec/columns/column_vector.cpp
@@ -86,12 +86,12 @@ void ColumnVector<T>::serialize_vec_with_null_map(std::vector<StringRef>& keys, 
             memcpy(dest, null_map + i, sizeof(UInt8));
             if (null_map[i] == 0) {
                 // If this row is not null, serialize the value
-                memcpy_fixed<T>(dest + 1, (char*)&data[i]);
-                keys[i].size += sizeof(T) + sizeof(UInt8);
-            } else {
-                // This row is null, do not serialize value
-                keys[i].size += sizeof(UInt8);
+                memcpy(dest + 1, (void*)&data[i], sizeof(T));
             }
+        }
+
+        for (size_t i = 0; i < num_rows; ++i) {
+            keys[i].size += sizeof(UInt8) + (1 - null_map[i]) * sizeof(T);
         }
     } else {
         // All rows are not null, serialize null & value
@@ -116,12 +116,22 @@ template <typename T>
 void ColumnVector<T>::deserialize_vec_with_null_map(std::vector<StringRef>& keys,
                                                     const size_t num_rows,
                                                     const uint8_t* null_map) {
-    for (size_t i = 0; i < num_rows; ++i) {
-        if (null_map[i] == 0) {
+    DCHECK(null_map != nullptr);
+    const bool has_null = simd::contain_byte(null_map, num_rows, 1);
+
+    if (has_null) {
+        for (size_t i = 0; i < num_rows; ++i) {
+            if (null_map[i] == 0) {
+                keys[i].data = deserialize_and_insert_from_arena(keys[i].data);
+                keys[i].size -= sizeof(T);
+            } else {
+                insert_default();
+            }
+        }
+    } else {
+        for (size_t i = 0; i < num_rows; ++i) {
             keys[i].data = deserialize_and_insert_from_arena(keys[i].data);
             keys[i].size -= sizeof(T);
-        } else {
-            insert_default();
         }
     }
 }