```sql
mysql> desc t_nullable;
+-------------+------+------+-------+---------+-------+
| Field       | Type | Null | Key   | Default | Extra |
+-------------+------+------+-------+---------+-------+
| m1          | INT  | Yes  | true  | NULL    |       |
| m2          | INT  | Yes  | true  | NULL    |       |
| key_string1 | TEXT | Yes  | false | NULL    | NONE  |
| key_string2 | TEXT | Yes  | false | NULL    | NONE  |
| key_string3 | TEXT | Yes  | false | NULL    | NONE  |
| key_int64_1 | INT  | Yes  | false | NULL    | NONE  |
| key_int64_2 | INT  | Yes  | false | NULL    | NONE  |
| key_int64_3 | INT  | Yes  | false | NULL    | NONE  |
| key_int64_4 | INT  | Yes  | false | NULL    | NONE  |
| key_int64_5 | INT  | Yes  | false | NULL    | NONE  |
+-------------+------+------+-------+---------+-------+
10 rows in set (0.00 sec)

mysql> select count(*) from t_nullable;
+------------+
| count(*)   |
+------------+
| 1000000000 |
+------------+
1 row in set (0.12 sec)
```
master
```sql
mysql> select  CAST((key_int64_1)  as  double),CAST((key_int64_2)  as  double),CAST((key_int64_3)  as  double),CAST((key_int64_4)  as  double),CAST((key_int64_5)  as  double),  min(m1)  from  t_nullable  group  by  CAST((key_int64_1)  as  double),CAST((key_int64_2)  as  double),CAST((key_int64_3)  as  double),CAST((key_int64_4)  as  double),CAST((key_int64_5)  as  double);
+-----------------------------+-----------------------------+-----------------------------+-----------------------------+-----------------------------+---------+
| cast(key_int64_1 as DOUBLE) | cast(key_int64_2 as DOUBLE) | cast(key_int64_3 as DOUBLE) | cast(key_int64_4 as DOUBLE) | cast(key_int64_5 as DOUBLE) | min(m1) |
+-----------------------------+-----------------------------+-----------------------------+-----------------------------+-----------------------------+---------+
|                           5 |                           6 |                           7 |                           8 |                           9 |       5 |
|                           2 |                           3 |                           4 |                           5 |                           6 |       2 |
|                           6 |                           7 |                           8 |                           9 |                          10 |       6 |
|                          10 |                          11 |                          12 |                          13 |                          14 |      10 |
|                           4 |                           5 |                           6 |                           7 |                           8 |       4 |
|                           1 |                           2 |                           3 |                           4 |                           5 |       1 |
|                           7 |                           8 |                           9 |                          10 |                          11 |       7 |
|                           9 |                          10 |                          11 |                          12 |                          13 |       9 |
|                           8 |                           9 |                          10 |                          11 |                          12 |       8 |
|                           3 |                           4 |                           5 |                           6 |                           7 |       3 |
+-----------------------------+-----------------------------+-----------------------------+-----------------------------+-----------------------------+---------+
10 rows in set (1.79 sec)

mysql> set parallel_pipeline_task_num=1;
Query OK, 0 rows affected (0.02 sec)

mysql> select  CAST((key_int64_1)  as  double),CAST((key_int64_2)  as  double),CAST((key_int64_3)  as  double),CAST((key_int64_4)  as  double),CAST((key_int64_5)  as  double),  min(m1)  from  t_nullable  group  by  CAST((key_int64_1)  as  double),CAST((key_int64_2)  as  double),CAST((key_int64_3)  as  double),CAST((key_int64_4)  as  double),CAST((key_int64_5)  as  double);
+-----------------------------+-----------------------------+-----------------------------+-----------------------------+-----------------------------+---------+
| cast(key_int64_1 as DOUBLE) | cast(key_int64_2 as DOUBLE) | cast(key_int64_3 as DOUBLE) | cast(key_int64_4 as DOUBLE) | cast(key_int64_5 as DOUBLE) | min(m1) |
+-----------------------------+-----------------------------+-----------------------------+-----------------------------+-----------------------------+---------+
|                           7 |                           8 |                           9 |                          10 |                          11 |       7 |
|                           5 |                           6 |                           7 |                           8 |                           9 |       5 |
|                           6 |                           7 |                           8 |                           9 |                          10 |       6 |
|                           1 |                           2 |                           3 |                           4 |                           5 |       1 |
|                           4 |                           5 |                           6 |                           7 |                           8 |       4 |
|                          10 |                          11 |                          12 |                          13 |                          14 |      10 |
|                           3 |                           4 |                           5 |                           6 |                           7 |       3 |
|                           2 |                           3 |                           4 |                           5 |                           6 |       2 |
|                           9 |                          10 |                          11 |                          12 |                          13 |       9 |
|                           8 |                           9 |                          10 |                          11 |                          12 |       8 |
+-----------------------------+-----------------------------+-----------------------------+-----------------------------+-----------------------------+---------+
10 rows in set (15.24 sec)
```


now
```sql
mysql> select  CAST((key_int64_1)  as  double),CAST((key_int64_2)  as  double),CAST((key_int64_3)  as  double),CAST((key_int64_4)  as  double),CAST((key_int64_5)  as  double),  min(m1)  from  t_nullable  group  by  CAST((key_int64_1)  as  double),CAST((key_int64_2)  as  double),CAST((key_int64_3)  as  double),CAST((key_int64_4)  as  double),CAST((key_int64_5)  as  double);
+-----------------------------+-----------------------------+-----------------------------+-----------------------------+-----------------------------+---------+
| cast(key_int64_1 as DOUBLE) | cast(key_int64_2 as DOUBLE) | cast(key_int64_3 as DOUBLE) | cast(key_int64_4 as DOUBLE) | cast(key_int64_5 as DOUBLE) | min(m1) |
+-----------------------------+-----------------------------+-----------------------------+-----------------------------+-----------------------------+---------+
|                           7 |                           8 |                           9 |                          10 |                          11 |       7 |
|                           1 |                           2 |                           3 |                           4 |                           5 |       1 |
|                           8 |                           9 |                          10 |                          11 |                          12 |       8 |
|                           6 |                           7 |                           8 |                           9 |                          10 |       6 |
|                           3 |                           4 |                           5 |                           6 |                           7 |       3 |
|                           5 |                           6 |                           7 |                           8 |                           9 |       5 |
|                           2 |                           3 |                           4 |                           5 |                           6 |       2 |
|                           4 |                           5 |                           6 |                           7 |                           8 |       4 |
|                          10 |                          11 |                          12 |                          13 |                          14 |      10 |
|                           9 |                          10 |                          11 |                          12 |                          13 |       9 |
+-----------------------------+-----------------------------+-----------------------------+-----------------------------+-----------------------------+---------+
10 rows in set (1.28 sec)
```
```
mysql> set parallel_pipeline_task_num=1;
Query OK, 0 rows affected (0.00 sec)
```
```sql
mysql> select  CAST((key_int64_1)  as  double),CAST((key_int64_2)  as  double),CAST((key_int64_3)  as  double),CAST((key_int64_4)  as  double),CAST((key_int64_5)  as  double),  min(m1)  from  t_nullable  group  by  CAST((key_int64_1)  as  double),CAST((key_int64_2)  as  double),CAST((key_int64_3)  as  double),CAST((key_int64_4)  as  double),CAST((key_int64_5)  as  double);
+-----------------------------+-----------------------------+-----------------------------+-----------------------------+-----------------------------+---------+
| cast(key_int64_1 as DOUBLE) | cast(key_int64_2 as DOUBLE) | cast(key_int64_3 as DOUBLE) | cast(key_int64_4 as DOUBLE) | cast(key_int64_5 as DOUBLE) | min(m1) |
+-----------------------------+-----------------------------+-----------------------------+-----------------------------+-----------------------------+---------+
|                           6 |                           7 |                           8 |                           9 |                          10 |       6 |
|                           5 |                           6 |                           7 |                           8 |                           9 |       5 |
|                           4 |                           5 |                           6 |                           7 |                           8 |       4 |
|                          10 |                          11 |                          12 |                          13 |                          14 |      10 |
|                           7 |                           8 |                           9 |                          10 |                          11 |       7 |
|                           1 |                           2 |                           3 |                           4 |                           5 |       1 |
|                           2 |                           3 |                           4 |                           5 |                           6 |       2 |
|                           3 |                           4 |                           5 |                           6 |                           7 |       3 |
|                           9 |                          10 |                          11 |                          12 |                          13 |       9 |
|                           8 |                           9 |                          10 |                          11 |                          12 |       8 |
+-----------------------------+-----------------------------+-----------------------------+-----------------------------+-----------------------------+---------+
10 rows in set (13.26 sec)
```

about 20% improvement