@zhiqiang-hhhh 2024-07-16
run buildall