@zhiqiang-hhhh 2024-07-17
run buildall